Refactor completo del proyecto a partir de versión antigua: mejoras en DTOs, tipos de datos y estructura de controladores
Este repositorio es un clon basado en una versión anterior del proyecto, sobre la cual se realizó una refactorización profunda.
Cambios principales:

    Renombrado de DTOs para mayor claridad, anteponiendo "Solicitud" y "Respuesta" según su responsabilidad (entrada/salida de datos).

    Cambio de tipo de datos BigDecimal a Long en atributos de ItemPedido y Pedido, para simplificar el manejo y mejorar la comprensión práctica del modelo.

    Reestructuración completa del controlador: ahora actúa como intermediario entre los DTOs y la capa de servicios, eliminando la lógica de negocio directamente acoplada.

    Reducción significativa del tamaño del código, haciéndolo más limpio, mantenible y alineado con buenas prácticas de arquitectura.

+++ src/main/java/com/ecomarket/pedidos/models/entity/ItemPedido.java
@@ -32,11 +32,11 @@ public class ItemPedido {
 
     @NotNull(message = "El precio al comprar no puede ser nulo.")
     @Column(nullable = false)
-    private Double precioAlComprar;
+    private long precioAlComprar;
 
     @NotNull(message = "El subtotal no puede ser nulo.")
     @Column(nullable = false)
-    private Double subTotal;
+    private long subTotal;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "pedido_id", nullable = false)
